refactor: move automatic bin size selection into  operation; add optional binning and smoothing to psth operation
