+++ src/aind_ephys_utils/ops/__init__.py
@@ -12,7 +12,6 @@ from .pseudopop import pseudopop
 from .psth import psth
 from .reduce import reduce
 from .restrict import restrict
-from .sliding_psth import sliding_psth
 from .smooth import smooth
 
 __all__ = [
@@ -24,6 +23,5 @@ __all__ = [
     "pseudopop",
     "reduce",
     "restrict",
-    "sliding_psth",
     "smooth",
 ]
